+++ src/main/java/pl/jojczykp/gumtree/address_book/Processor.java
@@ -4,7 +4,7 @@ import java.util.stream.Stream;
 
 public class Processor {
 
-	public long process(Stream<String> stream, Queries queries) {
+	public int process(Stream<String> stream, Queries queries) {
 		return stream
 				.map(Record::new)
 				.collect(queries.countingMales());
